massive refactoring and added agent
