+++ fulfilments/agent.py
@@ -1,0 +1,97 @@
+#!/usr/bin/env python
+import web
+
+from uuid import uuid4
+
+import logging
+from os import getenv
+from json import loads, dumps
+from pprint import pformat
+from requests import post
+from collections import defaultdict
+from os import _exit
+import signal
+from time import time
+from urllib import quote
+
+from simulation import SimulationDispatcher
+
+logging.basicConfig(level=logging.DEBUG)
+
+
+class SessionManager:
+    '''
+    A management class containing all active sessions
+    '''
+    def __init__(self):
+        self.__active_sessions = defaultdict(dict)
+
+    def get(self, session_id):
+        return self.__active_sessions[session_id]
+
+    def all(self):
+        return dict(self.__active_sessions)
+
+    def set(self, session_id, args={}):
+        self.__active_sessions[session_id] = args
+
+
+session_manager = SessionManager()
+
+
+class Agent:
+    '''
+    Class to send dialogflow event out via the standard API
+    '''
+    def __init__(self, apikey=None):
+        if apikey is not None:
+            self.apikey = apikey
+        else:
+            self.apikey = getenv('DF_APIKEY', 'invalidapikey')
+            logging.info('APIKEY: %s' % self.apikey)
+        self.header = {
+            'Authorization': 'Bearer %s' % self.apikey,
+            'Content-Type': 'application/json; charset=utf-8'
+        }
+        self.url = 'https://api.dialogflow.com/v1/query'
+
+    def query(self, q):
+        data = {
+            'contexts': [],
+            'query': q,
+            'lang': 'en',
+            'sessionId': str(uuid4())
+        }
+        r = post(self.url, data=dumps(data), headers=self.header)
+        fparams = r.json()
+        logging.info(pformat(fparams))
+        logging.info(pformat(SimulationDispatcher().dispatch(fparams)))
+        #logging.info(pformat(r.json()))
+
+    def send_event(self, event_name, parameters={}, sessionId=None):
+        if sessionId is not None:
+            return self._send_event(event_name, parameters, sessionId)
+        else:
+            return [
+                self._send_event(event_name, parameters, s)
+                for s in session_manager.all()
+            ]
+
+    def _send_event(self, event_name, parameters, sessionId):
+        logging.info('send event %s to session %s' % (event_name, sessionId))
+        data = {
+            'event': {
+                'name': event_name,
+                'data': parameters
+            },
+            'lang': 'en',
+            'sessionId': sessionId
+
+        }
+        r = post(self.url, data=dumps(data), headers=self.header)
+        logging.info(pformat(r.json()))
+
+
+if __name__ == "__main__":
+    agent = Agent()
+    agent.query("what is bielefeld?")
